refactor(api): move fromCalibrationScope function to Calibration model to switch dependence with shared
